refactor(telemetry): stop re-exporting recording from telemetry barrel
