+++ extension/src/extension/services/telemetry/index.ts
@@ -1,6 +1,3 @@
 // Telemetry services barrel exports
 export { TelemetryManager } from './telemetryManager';
 export * from './types';
-
-// Recording
-export * from './recording';
